refactor into generator and tests
