+++ package.json
@@ -5,9 +5,11 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.1.1",
     "bootstrap": "^5.1.3",
+    "fast-xml-parser": "^4.0.8",
     "fuse.js": "^6.6.2",
     "gray-matter": "^4.0.3",
-    "jquery": "^3.6.0"
+    "jquery": "^3.6.0",
+    "node-fetch": "^2.6.7"
   },
   "devDependencies": {
     "@playwright/test": "^1.20.1"
